Create semver package
* Bump the dev part to "dev.2"
* Add stubby setup.py file for compatibility with python 3.6
* Deprecate cli functions imported from root
* Revert to `pysemver` as console script.
* Refactor __main__.py
  * add type hints for correctness
* Rename VersionInfo to Version to close #305
* Refactor and integrate suggestion from @tomschr
  * Create :file:`src/semver/cli.py` for all CLI methods
  * Create :file:`src/semver/_deprecated.py` for the ``deprecated`` decorator and other deprecated functions
  * Create :file:`src/semver/__main__.py` to allow calling the CLI using :command:`python -m semver`
  * Create :file:`src/semver/_types.py` to hold type aliases
  * Create :file:`src/semver/version.py` to hold the :class:`VersionInfo` class and its utility functions
  * Create :file:`src/semver/__about__.py` for all the metadata variables
* Adapt infrastructure code to the new project layout.
  * Replace :file:`setup.py` with :file:`setup.cfg` because the :file:`setup.cfg` is easier to use
  * Adapt documentation code snippets where needed
  * Adapt tests
  * Changed the ``deprecated`` to hardcode the ``semver`` package name in the warning.
* Change path for docformatter and run it.
* Remove pyi inclusion from black sine we aren't using them
* Split up changelog to make more sense.
* Add documentation for Version rename
* Increase coverage to 100% in non-deprecated parts.
* Update changelog.d/169.feature.rst
