+++ changelog.d/169.deprecation.rst
@@ -1,0 +1,1 @@
+Deprecate CLI functions not imported from ``semver.cli``.
