+++ changelog.d/305.doc.rst
@@ -1,0 +1,1 @@
+Add note about :class:`Version` rename.
