+++ changelog.d/305.feature.rst
@@ -1,0 +1,1 @@
+Rename :class:`VersionInfo` to :class:`Version` but keep an alias for compatibility
