+++ setup.cfg
@@ -1,3 +1,46 @@
+[metadata]
+name = semver
+version = attr: semver.__about__.__version__
+description = attr: semver.__about__.__description__
+long_description = file: README.rst
+author = attr: semver.__about__.__author__
+author_email = attr: semver.__about__.__author_email__
+maintainer = attr: semver.__about__.__maintainer__
+maintainer_email = attr: semver.__about__.__maintainer_email__
+url = https://github.com/python-semver/python-semver
+download_url = https://github.com/python-semver/python-semver/downloads
+project_urls =
+    Documentation = https://python-semver.rtfd.io
+    Releases = https://github.com/python-semver/python-semver/releases
+    Bug Tracker = https://github.com/python-semver/python-semver/issues
+classifiers =
+    Environment :: Web Environment
+    Intended Audience :: Developers
+    License :: OSI Approved :: BSD License
+    Operating System :: OS Independent
+    Programming Language :: Python
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3.6
+    Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
+    Topic :: Software Development :: Libraries :: Python Modules
+license = BSD
+
+[options]
+package_dir =
+    =src
+packages = find:
+python_requires = >=3.6.*
+include_package_data = True
+
+[options.entry_points]
+console_scripts =
+    pysemver = semver.cli:main
+
+[options.packages.find]
+where = src
+
 [tool:pytest]
 norecursedirs = .git build .env/ env/ .pyenv/ .tmp/ .eggs/ venv/
 testpaths = tests docs
@@ -15,13 +58,14 @@ addopts =
 max-line-length = 88
 ignore = F821,W503
 exclude =
-    .env,
-    venv,
-    .eggs,
-    .tox,
-    .git,
-    __pycache__,
-    build,
+    src/semver/__init__.py
+    .env
+    venv
+    .eggs
+    .tox
+    .git
+    __pycache__
+    build
     dist
     docs
     conftest.py
@@ -32,6 +76,7 @@ count = False
 max-line-length = 88
 statistics = True
 exclude =
+    src/semver/__init__.py
     .env,
     .eggs,
     .tox,
