+++ src/semver/__about__.py
@@ -1,0 +1,8 @@
+__version__ = "3.0.0-dev.2"
+__author__ = "Kostiantyn Rybnikov"
+__author_email__ = "k-bx@k-bx.com"
+__maintainer__ = ["Sebastien Celles", "Tom Schraitle"]
+__maintainer_email__ = "s.celles@gmail.com"
+__description__ = "Python helper for Semantic Versioning (http://semver.org)"
+
+SEMVER_SPEC_VERSION = "2.0.0"
