+++ tests/conftest.py
@@ -23,8 +23,8 @@ def version():
     Creates a version
 
     :return: a version type
-    :rtype: VersionInfo
+    :rtype: Version
     """
-    return semver.VersionInfo(
+    return semver.Version(
         major=1, minor=2, patch=3, prerelease="alpha.1.2", build="build.11.e0f985a"
     )
