+++ tests/test_replace.py
@@ -1,6 +1,6 @@
 import pytest
 
-from semver import VersionInfo, replace
+from semver import Version, replace
 
 
 @pytest.mark.parametrize(
@@ -42,9 +42,9 @@ def test_replace_raises_TypeError_for_invalid_keyword_arg():
     ],
 )
 def test_should_return_versioninfo_with_replaced_parts(version, parts, expected):
-    assert VersionInfo.parse(version).replace(**parts) == VersionInfo.parse(expected)
+    assert Version.parse(version).replace(**parts) == Version.parse(expected)
 
 
 def test_replace_raises_ValueError_for_non_numeric_values():
     with pytest.raises(ValueError):
-        VersionInfo.parse("1.2.3").replace(major="x")
+        Version.parse("1.2.3").replace(major="x")
