+++ tests/test_semver.py
@@ -1,13 +1,13 @@
 import pytest  # noqa
 
-from semver import VersionInfo
+from semver import Version
 
 
 @pytest.mark.parametrize(
     "string,expected", [("rc", "rc"), ("rc.1", "rc.2"), ("2x", "3x")]
 )
 def test_should_private_increment_string(string, expected):
-    assert VersionInfo._increment_string(string) == expected
+    assert Version._increment_string(string) == expected
 
 
 @pytest.mark.parametrize(
@@ -21,7 +21,7 @@ def test_should_private_increment_string(string, expected):
 )
 def test_should_not_allow_negative_numbers(ver):
     with pytest.raises(ValueError, match=".* is negative. .*"):
-        VersionInfo(**ver)
+        Version(**ver)
 
 
 def test_should_versioninfo_to_dict(version):
@@ -47,31 +47,36 @@ def test_version_info_should_be_iterable(version):
 
 
 def test_should_be_able_to_use_strings_as_major_minor_patch():
-    v = VersionInfo("1", "2", "3")
+    v = Version("1", "2", "3")
     assert isinstance(v.major, int)
     assert isinstance(v.minor, int)
     assert isinstance(v.patch, int)
     assert v.prerelease is None
     assert v.build is None
-    assert VersionInfo("1", "2", "3") == VersionInfo(1, 2, 3)
+    assert Version("1", "2", "3") == Version(1, 2, 3)
 
 
 def test_using_non_numeric_string_as_major_minor_patch_throws():
     with pytest.raises(ValueError):
-        VersionInfo("a")
+        Version("a")
     with pytest.raises(ValueError):
-        VersionInfo(1, "a")
+        Version(1, "a")
     with pytest.raises(ValueError):
-        VersionInfo(1, 2, "a")
+        Version(1, 2, "a")
 
 
 def test_should_be_able_to_use_integers_as_prerelease_build():
-    v = VersionInfo(1, 2, 3, 4, 5)
+    v = Version(1, 2, 3, 4, 5)
     assert isinstance(v.prerelease, str)
     assert isinstance(v.build, str)
-    assert VersionInfo(1, 2, 3, 4, 5) == VersionInfo(1, 2, 3, "4", "5")
+    assert Version(1, 2, 3, 4, 5) == Version(1, 2, 3, "4", "5")
 
 
 def test_should_versioninfo_isvalid():
-    assert VersionInfo.isvalid("1.0.0") is True
-    assert VersionInfo.isvalid("foo") is False
+    assert Version.isvalid("1.0.0") is True
+    assert Version.isvalid("foo") is False
+
+
+def test_versioninfo_compare_should_raise_when_passed_invalid_value():
+    with pytest.raises(TypeError):
+        Version(1, 2, 3).compare(4)
